added function 14 to scan database names
This function has the capability to scan the domains folder for all domains and print out the DB_NAME of each website if it's a WordPress application. It can also allow the user to submit a domain name to scan, then scan all folders inside the main domain to check if there is a subdomain folder with WordPress installed, and if yes, it'll print out all the databases for the subdomains as well. It can be used both from public_html and also from the domains/ tab on shared hosting plans on hPanel